Small adjustments for moved commands
